Face recognition web
This web application harnesses the power of the Clairfi Face Recognition API to provide accurate and real-time face detection and recognition within images. Users can upload images, and the application seamlessly interacts with the API to identify faces and display the results. This web make the combination of JavaScript, React and the Clairfi API.